+++ tests/integration/compose-up.sh
@@ -22,61 +22,22 @@
 #  along with Slurm-Mail.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-function catch {
-  if [ "$1" != "0" ]; then
-    #echo "Error $1 occurred on line $2" 1>&2
-    tidyup $NAME
-  fi
-}
-
-function tidyup {
-  if [ $KEEP_CONTAINER -eq 0 ]; then
-    echo "stopping container..."
-    docker container stop $1
-    echo "deleting container..."
-    docker container rm $1
-    echo "done"
-  fi
-}
+set -e
 
 function usage {
-  echo "Usage: $0 -s SLURM_VERSION [-k] [-m] [-r] [-t TEST_NAME] [-v]" 1>&2
-  echo "  -k                   keep the test container upon failure"
-  echo "  -m                   show e-mail log"
+  echo "Usage: $0 -s SLURM_VERSION [-r]" 1>&2
   echo "  -r                   don't build slurm-mail RPM - use existing file"
   echo "  -s SLURM_VERSION     version of Slurm to test against"
-  echo "  -t TEST_NAME         only run this named test"
-  echo "  -v                   turn on debugging"
   exit 0
 }
 
-set -e
-trap 'catch $? $LINENO' EXIT
-
-KEEP_CONTAINER=0
-MAIL_LOG=0
-USE_RPM=0
-VERBOSE=0
-
-while getopts ":kms:rt:v" options; do
+while getopts ":s:r" options; do
   case "${options}" in
-    k)
-      KEEP_CONTAINER=1
-      ;;
-    m)
-      MAIL_LOG=1
-      ;;
     r)
       USE_RPM=1
       ;;
     s)
       SLURM_VER=${OPTARG}
-      ;;
-    t)
-      RUN_TEST=${OPTARG}
-      ;;
-    v)
-      VERBOSE=1
       ;;
     :)
       echo "Error: -${OPTARG} requires a value"
@@ -90,19 +51,6 @@ done
 
 if [ -z $SLURM_VER ]; then
   usage
-fi
-
-OPTS=""
-if [ ! -z $RUN_TEST ]; then
-  OPTS="-t $RUN_TEST"
-fi
-
-if [ $VERBOSE -eq 1 ]; then
-  OPTS="$OPTS -v"
-fi
-
-if [ $MAIL_LOG -eq 1 ]; then
-  OPTS="$OPTS -m"
 fi
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
@@ -122,15 +70,5 @@ fi
 cd $DIR
 RPM=`ls -1 slurm-mail*.rpm`
 
-docker build \
-  --build-arg DISABLE_CRON=1 \
-  --build-arg SLURM_MAIL_RPM=${RPM} \
-  --build-arg SLURM_VER=${SLURM_VER} \
-  -t neilmunday/slurm-mail:${SLURM_VER} \
-  -f Dockerfile.slurm-mail .
-
-docker run -d -h compute --name $NAME neilmunday/slurm-mail:${SLURM_VER}
-
-docker exec $NAME /bin/bash -c "/root/testing/run-tests.py -i /root/testing/tests.yml -o /root/testing/output $OPTS"
-
-tidyup $NAME
+docker compose build --build-arg SLURM_VER=$SLURM_VER --build-arg SLURM_MAIL_RPM=$RPM
+docker compose up